Refactor `@lock` endpoint based on CRUD operations - refs #1181
