+++ news/1181.feature
@@ -1,0 +1,1 @@
+Refactor `@lock` endpoint based on CRUD operations and add it to expandable `@components` [@avoinea]
