+++ src/plone/restapi/services/content/update.py
@@ -2,7 +2,7 @@ from plone.restapi.exceptions import DeserializationError
 from plone.restapi.interfaces import IDeserializeFromJson
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
-from plone.restapi.services.locking.locking import is_locked
+from plone.restapi.services.locking import is_locked
 from zope.component import queryMultiAdapter
