+++ src/plone/restapi/services/locking/get.py
@@ -1,0 +1,9 @@
+from plone.restapi.services import Service
+from plone.restapi.services.locking import lock_info
+
+
+class Lock(Service):
+    """Lock information about the current lock"""
+
+    def reply(self):
+        return lock_info(self.context)
