+++ src/plone/restapi/tests/http-examples/refresh_lock.req
@@ -1,3 +1,3 @@
-POST /plone/front-page/@refresh-lock HTTP/1.1
+PATCH /plone/front-page/@lock HTTP/1.1
 Accept: application/json
 Authorization: Basic YWRtaW46c2VjcmV0
