+++ src/plone/restapi/tests/http-examples/types_document.resp
@@ -97,7 +97,7 @@ Content-Type: application/json+schema
       }
     }, 
     "author_email": {
-      "behavior": "plone.dexterity.schema.generated.plone_0_Document", 
+      "behavior": "plone.dexterity.schema.generated.plone_5_1626959789_2_0922327_0_Document", 
       "description": "Email of the author", 
       "factory": "Email", 
       "title": "Author email",
