+++ src/plone/restapi/tests/http-examples/types_document_post_field.resp
@@ -2,7 +2,7 @@ HTTP/1.1 201 Created
 Content-Type: application/json
 
 {
-  "behavior": "plone.dexterity.schema.generated.plone_0_Document", 
+  "behavior": "plone.dexterity.schema.generated.plone_5_1626959789_2_0811691_0_Document", 
   "description": "Email of the author", 
   "factory": "Email", 
   "title": "Author email",
