+++ src/plone/restapi/tests/http-examples/types_document_put.req
@@ -54,7 +54,7 @@ Content-Type: application/json
             }
         },
         "author_email": {
-            "behavior": "plone.dexterity.schema.generated.plone_0_Document",
+            "behavior": "plone.dexterity.schema.generated.plone_5_1626959789_2_0922327_0_Document",
             "description": "Email of the author",
             "factory": "Email",
             "title": "Author email",
