+++ src/plone/restapi/tests/http-examples/unlock.req
@@ -1,3 +1,3 @@
-POST /plone/front-page/@unlock HTTP/1.1
+DELETE /plone/front-page/@lock HTTP/1.1
 Accept: application/json
 Authorization: Basic YWRtaW46c2VjcmV0
